requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.3.1
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.29.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h5py==3.10.0
idna==3.6
Jinja2==3.1.3
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pocketsphinx==5.0.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
pypiwin32==223
PySoundFile==0.9.0.post1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
pywin32==306
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
setuptools==69.2.0
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.1
srt==3.5.3
sympy==1.12
termcolor==2.4.0
tiktoken==0.6.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vosk==0.3.45
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0